YUUUUUHHHHHHHHHHHHHHHHHHHHHHH
BROOOOOOOOOOOOOOO

YYEEEEEEESSSSSSSSSSSSS

WWWWWWWWWWWWWWWWW

Ok we're getting somewhere. Got the rest of the 15 mirrorless cameras into an extracted json format, just working on mapping the schema to my current postgres layout. but the pipelining is actually working FINALLYYYYYYY

I'm hoping i'm not jinxing myself on this one, but should be a tad easier to get this ball rolling at least with mirrorless cameras for the rest of the stats......... for now (until pdf reads & getting info from non-official sources lol)